+++ db/users_db.py
@@ -1,13 +1,13 @@
 from tokenize import Floatnumber
-from sqlalchemy import Column, Integer, String
+from sqlalchemy import Column, Integer, Float, String
 from db.db_connection import Base, engine
 
 class UserInDB(Base):
     __tablename__="users"
     username    = Column(String, primary_key=True, unique=True)
     password    = Column(String)
-    liabilities = Column(Integer)
-    passives    = Column(Integer)
+    liabilities = Column(Float)
+    passives    = Column(Float)
     
 
 Base.metadata.create_all(bind=engine)
